requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-jose[cryptography]==3.4.0
httpx==0.27.2
python-dotenv==1.0.1
